% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagesurvey.R
\docType{package}
\name{packagesurvey}
\alias{packagesurvey}
\alias{packagesurvey-package}
\alias{packagesurvey-package}
\title{Survey of R users on package discovery}
\description{
This package contains results from an online survey run in preparation for
a session at useR!2017 focused on discovering, learning about, and evaluting
R packages.
}
